autodoc-swarm/pyproject.toml
[project]
name = "autodoc-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.4.12",
    "langchain-anthropic>=1.4.0",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "tenacity>=9.1.4",
    "typer>=0.24.1",
]
